Fix float conversion in icon bounce animation

+++ app/src/main/java/com/termux/app/SuggestionBarView.java
@@ -3532,7 +3532,7 @@ public final class SuggestionBarView extends GridLayout {
         sourceView.animate().cancel();
         sourceView.setPivotX(sourceView.getWidth() * 0.5f);
         sourceView.setPivotY(sourceView.getHeight());
-        float lift = dp(4.2f);
+        float lift = (float) dp(4.2f);
         sourceView.animate()
             .translationY(-lift)
             .scaleX(1.08f)
@@ -3554,7 +3554,7 @@ public final class SuggestionBarView extends GridLayout {
         final float startTranslationY = sourceView.getTranslationY();
         final float startScaleX = sourceView.getScaleX();
         final float startScaleY = sourceView.getScaleY();
-        final float lift = dp(4.2f);
+        final float lift = (float) dp(4.2f);
         ValueAnimator animator = ValueAnimator.ofFloat(0f, 1f);
         animator.setDuration(760L);
         animator.setInterpolator(new LinearInterpolator());
